Use a forked child process to make Bundler calls that irreversibly modify the Ruby state
When using a Git repository as a dependency, Bundler loads its .gemspec file, which in turn may modify the Ruby state arbitrarily in ways that aren't readily reversible. To sidestep such behavior, the plugin now forks a child process for making sensitive Bundler calls. This should resolve issue #14, which exemplifies the arbitrary badness that can result from not being able to "roll back" the Ruby state after inspecting a Gemfile.